refactor: context의 state들을 custom hook으로 분리, redcuer 분리
